requirement.txt
mysql-connector-python
psycopg2-binary
pymongo
Faker